requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
Authlib==1.3.0
bidict==0.23.1
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
chainlit==1.0.301
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
cryptography==42.0.5
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
dotenv-python==0.0.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.108.0
fastapi-socketio==0.0.10
filelock==3.13.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-health-checking==1.62.0
grpcio-tools==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
huggingface-hub==0.21.1
idna==3.6
importlib-metadata==6.11.0
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langgraph==0.0.26
langsmith==0.1.9
Lazify==0.4.0
literalai==0.0.204
llama_cpp_python==0.2.52
MarkupSafe==2.1.5
marshmallow==3.21.0
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
ollama==0.1.6
openai==1.12.0
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
orjson==3.9.15
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.11.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.4.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.32.0.post1
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
triton==2.2.0
typing-inspect==0.9.0
typing_extensions==4.10.0
uptrace==1.22.0
urllib3==2.2.1
uvicorn==0.25.0
validators==0.22.0
watchfiles==0.20.0
wcwidth==0.2.13
weaviate-client==4.4.4
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
